requirements.txt
django==2.0.2
django-uuid-upload
djangorestframework==3.7.7
django-filter==1.1.0
requests
mock
django-cors-headers
django-nose
coverage